requirements.txt
django==5.0.4
psycopg2==2.9.9
django-environ==0.11.2
pytest-django==4.8.0
django-ninja==1.1.0
ruff==0.2.2
mypy==1.9.0
pre-commit==3.7.0